requirements.txt
# ============================================
# StoryProof - Python Dependencies
# ============================================

# Core Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication & Security
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# AI & Machine Learning
google-genai>=0.2.0
google-api-core>=2.15.0
sentence-transformers>=2.3.0
pinecone-client>=3.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
chardet>=5.2.0
tqdm>=4.66.0

# Background Tasks
celery[redis]>=5.3.0
redis>=5.0.0
